requirements.txt
loguru==0.7.3
mcp==1.6.0
openai==1.74.0
protobuf==5.29.3
protobuf3-to-dict==0.1.5
PyExecJS==1.5.1
PyMySQL==1.1.0
python-dotenv==1.1.0
requests==2.31.0
SQLAlchemy==2.0.30
websockets==15.0.1
fastmcp==2.1.2
Brotli==1.1.0
cryptography==44.0.2